Introduce ContentPage component and organize pages
