+++ takwimu/takwimu_ui/public/index.html
@@ -156,51 +156,62 @@
           }
         },
         services: {
-          overview: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
+          overview:
+            'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
               tempor incididunt ut labore et dolore magna aliqua. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
               tempor incididunt ut labore et dolore magna aliqua',
 
-          servicesList: [{ 'title': 'Service one name',
-            'description': 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
+          servicesList: [
+            {
+              title: 'Service one name',
+              description:
+                'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
               tempor incididunt ut labore et dolore magna aliqua.'
-          }]
+            }
+          ]
         },
         faqs: {
-          overview: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo <a>Get in touch with us</a>',
-          faqsList: [{ 'question': 'Service Where is this data sourced from?',
-            'answer': 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
+          overview:
+            'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo <a>Get in touch with us</a>',
+          faqsList: [
+            {
+              question: 'Service Where is this data sourced from?',
+              answer:
+                'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
               tempor incididunt ut labore et dolore magna aliqua.'
-          }]
+            }
+          ]
         },
-        methodology: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
+        methodology:
+          'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
               tempor incididunt ut labore et dolore magna aliqua. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmo \
               tempor incididunt ut labore et dolore magna aliqu',
         relatedContent: [
-        {
-          type: "link",
-          value: {
-            type: "url",
+          {
+            type: 'link',
             value: {
-              title: "sajdfkldshkjgfs",
-              link: "http://takwimu.africa"
+              type: 'url',
+              value: {
+                title: 'sajdfkldshkjgfs',
+                link: 'http://takwimu.africa'
+              },
+              id: 'de517671-eeba-4c8d-990e-1c7496acae1c'
             },
-            id: "de517671-eeba-4c8d-990e-1c7496acae1c"
+            id: 'aa08e9eb-b54c-40e9-9339-5185d1a1d4fe'
           },
-          id: "aa08e9eb-b54c-40e9-9339-5185d1a1d4fe"
-        },
-        {
-          type: "link",
-          value: {
-            type: "url",
+          {
+            type: 'link',
             value: {
-              title: "kjajglkfsjlkaajkgagd;lksadgsa",
-              link: "http://instagram.com"
+              type: 'url',
+              value: {
+                title: 'kjajglkfsjlkaajkgagd;lksadgsa',
+                link: 'http://instagram.com'
+              },
+              id: '5f3786b3-b639-46bc-b6cb-9147fa52121d'
             },
-            id: "5f3786b3-b639-46bc-b6cb-9147fa52121d"
-          },
-          id: "4e87f44d-af08-459f-9757-b49bb69897b7"
+            id: '4e87f44d-af08-459f-9757-b49bb69897b7'
           }
-        ],
+        ]
       };
       var profileData = {
         demographics: {},
@@ -250,27 +261,20 @@
         page = 'analysis';
       }
       takwimu.page.id = page;
+      let ids = [];
       switch (page) {
         case 'about':
-          const elm = document.createElement('div');
-          elm.setAttribute('id', 'takwimuAboutUsPage');
-          body.appendChild(elm);
+          ids = ['takwimuAboutUsPage'];
           break;
         case 'topic':
-          ['takwimuProfileDetail', 'takwimuProfile'].forEach(id => {
-            const el = document.createElement('div');
-            el.setAttribute('id', id);
-            body.appendChild(el);
-          });
+          ids = ['takwimuProfileDetail', 'takwimuProfile'];
           break;
         case 'analysis':
-          const el = document.createElement('div');
-          el.setAttribute('id', 'takwimuAnalysisPage');
-          body.appendChild(el);
+          ids = ['takwimuAnalysisPage'];
           break;
         case 'homepage': // fall through
         default:
-          [
+          ids = [
             'takwimuHero',
             'takwimuFeaturedAnalysis',
             'takwimuFeaturedData',
@@ -278,12 +282,13 @@
             'takwimuMakingOf',
             'takwimuLatestNewsStories',
             'takwimuWhereToNext'
-          ].forEach(id => {
-            const el = document.createElement('div');
-            el.setAttribute('id', id);
-            body.appendChild(el);
-          });
+          ];
       }
+      ids.forEach(id => {
+        const el = document.createElement('div');
+        el.setAttribute('id', id);
+        body.appendChild(el);
+      });
     </script>
     <div id="takwimuFooter"></div>
